+++ src/main/java/kitchenpos/eatinorders/domain/JpaOrderRepository.java
@@ -1,5 +1,6 @@
 package kitchenpos.eatinorders.domain;
 
+import kitchenpos.eatinorders.tobe.domain.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
