+++ src/main/java/kitchenpos/eatinorders/domain/Order.java
@@ -1,115 +1,115 @@
-package kitchenpos.eatinorders.domain;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.UUID;
-
-@Table(name = "orders")
-@Entity
-public class Order {
-    @Column(name = "id", columnDefinition = "binary(16)")
-    @Id
-    private UUID id;
-
-    @Column(name = "type", nullable = false)
-    @Enumerated(EnumType.STRING)
-    private OrderType type;
-
-    @Column(name = "status", nullable = false)
-    @Enumerated(EnumType.STRING)
-    private OrderStatus status;
-
-    @Column(name = "order_date_time", nullable = false)
-    private LocalDateTime orderDateTime;
-
-    @OneToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE})
-    @JoinColumn(
-        name = "order_id",
-        nullable = false,
-        columnDefinition = "binary(16)",
-        foreignKey = @ForeignKey(name = "fk_order_line_item_to_orders")
-    )
-    private List<OrderLineItem> orderLineItems;
-
-    @Column(name = "delivery_address")
-    private String deliveryAddress;
-
-    @ManyToOne
-    @JoinColumn(
-        name = "order_table_id",
-        columnDefinition = "binary(16)",
-        foreignKey = @ForeignKey(name = "fk_orders_to_order_table")
-    )
-    private OrderTable orderTable;
-
-    @Transient
-    private UUID orderTableId;
-
-    public Order() {
-    }
-
-    public UUID getId() {
-        return id;
-    }
-
-    public void setId(final UUID id) {
-        this.id = id;
-    }
-
-    public OrderType getType() {
-        return type;
-    }
-
-    public void setType(final OrderType type) {
-        this.type = type;
-    }
-
-    public OrderStatus getStatus() {
-        return status;
-    }
-
-    public void setStatus(final OrderStatus status) {
-        this.status = status;
-    }
-
-    public LocalDateTime getOrderDateTime() {
-        return orderDateTime;
-    }
-
-    public void setOrderDateTime(final LocalDateTime orderDateTime) {
-        this.orderDateTime = orderDateTime;
-    }
-
-    public List<OrderLineItem> getOrderLineItems() {
-        return orderLineItems;
-    }
-
-    public void setOrderLineItems(final List<OrderLineItem> orderLineItems) {
-        this.orderLineItems = orderLineItems;
-    }
-
-    public String getDeliveryAddress() {
-        return deliveryAddress;
-    }
-
-    public void setDeliveryAddress(final String deliveryAddress) {
-        this.deliveryAddress = deliveryAddress;
-    }
-
-    public OrderTable getOrderTable() {
-        return orderTable;
-    }
-
-    public void setOrderTable(final OrderTable orderTable) {
-        this.orderTable = orderTable;
-    }
-
-    public UUID getOrderTableId() {
-        return orderTableId;
-    }
-
-    public void setOrderTableId(final UUID orderTableId) {
-        this.orderTableId = orderTableId;
-    }
-}
+//package kitchenpos.eatinorders.domain;
+//
+//import javax.persistence.*;
+//import java.time.LocalDateTime;
+//import java.util.List;
+//import java.util.UUID;
+//
+//@Table(name = "orders")
+//@Entity
+//public class Order {
+//    @Column(name = "id", columnDefinition = "binary(16)")
+//    @Id
+//    private UUID id;
+//
+//    @Column(name = "type", nullable = false)
+//    @Enumerated(EnumType.STRING)
+//    private OrderType type;
+//
+//    @Column(name = "status", nullable = false)
+//    @Enumerated(EnumType.STRING)
+//    private OrderStatus status;
+//
+//    @Column(name = "order_date_time", nullable = false)
+//    private LocalDateTime orderDateTime;
+//
+//    @OneToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE})
+//    @JoinColumn(
+//        name = "order_id",
+//        nullable = false,
+//        columnDefinition = "binary(16)",
+//        foreignKey = @ForeignKey(name = "fk_order_line_item_to_orders")
+//    )
+//    private List<OrderLineItem> orderLineItems;
+//
+//    @Column(name = "delivery_address")
+//    private String deliveryAddress;
+//
+//    @ManyToOne
+//    @JoinColumn(
+//        name = "order_table_id",
+//        columnDefinition = "binary(16)",
+//        foreignKey = @ForeignKey(name = "fk_orders_to_order_table")
+//    )
+//    private OrderTable orderTable;
+//
+//    @Transient
+//    private UUID orderTableId;
+//
+//    public Order() {
+//    }
+//
+//    public UUID getId() {
+//        return id;
+//    }
+//
+//    public void setId(final UUID id) {
+//        this.id = id;
+//    }
+//
+//    public OrderType getType() {
+//        return type;
+//    }
+//
+//    public void setType(final OrderType type) {
+//        this.type = type;
+//    }
+//
+//    public OrderStatus getStatus() {
+//        return status;
+//    }
+//
+//    public void setStatus(final OrderStatus status) {
+//        this.status = status;
+//    }
+//
+//    public LocalDateTime getOrderDateTime() {
+//        return orderDateTime;
+//    }
+//
+//    public void setOrderDateTime(final LocalDateTime orderDateTime) {
+//        this.orderDateTime = orderDateTime;
+//    }
+//
+//    public List<OrderLineItem> getOrderLineItems() {
+//        return orderLineItems;
+//    }
+//
+//    public void setOrderLineItems(final List<OrderLineItem> orderLineItems) {
+//        this.orderLineItems = orderLineItems;
+//    }
+//
+//    public String getDeliveryAddress() {
+//        return deliveryAddress;
+//    }
+//
+//    public void setDeliveryAddress(final String deliveryAddress) {
+//        this.deliveryAddress = deliveryAddress;
+//    }
+//
+//    public OrderTable getOrderTable() {
+//        return orderTable;
+//    }
+//
+//    public void setOrderTable(final OrderTable orderTable) {
+//        this.orderTable = orderTable;
+//    }
+//
+//    public UUID getOrderTableId() {
+//        return orderTableId;
+//    }
+//
+//    public void setOrderTableId(final UUID orderTableId) {
+//        this.orderTableId = orderTableId;
+//    }
+//}
