+++ src/main/java/kitchenpos/eatinorders/domain/OrderLineItem.java
@@ -1,85 +1,85 @@
-package kitchenpos.eatinorders.domain;
-
-import kitchenpos.menus.tobe.domain.Menu;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.ForeignKey;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-import javax.persistence.Transient;
-import java.math.BigDecimal;
-import java.util.UUID;
-
-@Table(name = "order_line_item")
-@Entity
-public class OrderLineItem {
-    @Column(name = "seq")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Id
-    private Long seq;
-
-    @ManyToOne(optional = false)
-    @JoinColumn(
-            name = "menu_id",
-            columnDefinition = "binary(16)",
-            foreignKey = @ForeignKey(name = "fk_order_line_item_to_menu")
-    )
-    private Menu menu;
-
-    @Column(name = "quantity", nullable = false)
-    private long quantity;
-
-    @Transient
-    private UUID menuId;
-
-    @Transient
-    private BigDecimal price;
-
-    public OrderLineItem() {
-    }
-
-    public Long getSeq() {
-        return seq;
-    }
-
-    public void setSeq(final Long seq) {
-        this.seq = seq;
-    }
-
-    public Menu getMenu() {
-        return menu;
-    }
-
-    public void setMenu(final Menu menu) {
-        this.menu = menu;
-    }
-
-    public long getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(final long quantity) {
-        this.quantity = quantity;
-    }
-
-    public UUID getMenuId() {
-        return menuId;
-    }
-
-    public void setMenuId(final UUID menuId) {
-        this.menuId = menuId;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
-    }
-
-    public void setPrice(final BigDecimal price) {
-        this.price = price;
-    }
-}
+//package kitchenpos.eatinorders.domain;
+//
+//import kitchenpos.menus.tobe.domain.Menu;
+//
+//import javax.persistence.Column;
+//import javax.persistence.Entity;
+//import javax.persistence.ForeignKey;
+//import javax.persistence.GeneratedValue;
+//import javax.persistence.GenerationType;
+//import javax.persistence.Id;
+//import javax.persistence.JoinColumn;
+//import javax.persistence.ManyToOne;
+//import javax.persistence.Table;
+//import javax.persistence.Transient;
+//import java.math.BigDecimal;
+//import java.util.UUID;
+//
+//@Table(name = "order_line_item")
+//@Entity
+//public class OrderLineItem {
+//    @Column(name = "seq")
+//    @GeneratedValue(strategy = GenerationType.IDENTITY)
+//    @Id
+//    private Long seq;
+//
+//    @ManyToOne(optional = false)
+//    @JoinColumn(
+//            name = "menu_id",
+//            columnDefinition = "binary(16)",
+//            foreignKey = @ForeignKey(name = "fk_order_line_item_to_menu")
+//    )
+//    private Menu menu;
+//
+//    @Column(name = "quantity", nullable = false)
+//    private long quantity;
+//
+//    @Transient
+//    private UUID menuId;
+//
+//    @Transient
+//    private BigDecimal price;
+//
+//    public OrderLineItem() {
+//    }
+//
+//    public Long getSeq() {
+//        return seq;
+//    }
+//
+//    public void setSeq(final Long seq) {
+//        this.seq = seq;
+//    }
+//
+//    public Menu getMenu() {
+//        return menu;
+//    }
+//
+//    public void setMenu(final Menu menu) {
+//        this.menu = menu;
+//    }
+//
+//    public long getQuantity() {
+//        return quantity;
+//    }
+//
+//    public void setQuantity(final long quantity) {
+//        this.quantity = quantity;
+//    }
+//
+//    public UUID getMenuId() {
+//        return menuId;
+//    }
+//
+//    public void setMenuId(final UUID menuId) {
+//        this.menuId = menuId;
+//    }
+//
+//    public BigDecimal getPrice() {
+//        return price;
+//    }
+//
+//    public void setPrice(final BigDecimal price) {
+//        this.price = price;
+//    }
+//}
