+++ src/main/java/kitchenpos/eatinorders/domain/OrderTable.java
@@ -1,59 +1,59 @@
-package kitchenpos.eatinorders.domain;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import java.util.UUID;
-
-@Table(name = "order_table")
-@Entity
-public class OrderTable {
-    @Column(name = "id", columnDefinition = "binary(16)")
-    @Id
-    private UUID id;
-
-    @Column(name = "name", nullable = false)
-    private String name;
-
-    @Column(name = "number_of_guests", nullable = false)
-    private int numberOfGuests;
-
-    @Column(name = "occupied", nullable = false)
-    private boolean occupied;
-
-    public OrderTable() {
-    }
-
-    public UUID getId() {
-        return id;
-    }
-
-    public void setId(final UUID id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(final String name) {
-        this.name = name;
-    }
-
-    public int getNumberOfGuests() {
-        return numberOfGuests;
-    }
-
-    public void setNumberOfGuests(final int numberOfGuests) {
-        this.numberOfGuests = numberOfGuests;
-    }
-
-    public boolean isOccupied() {
-        return occupied;
-    }
-
-    public void setOccupied(final boolean occupied) {
-        this.occupied = occupied;
-    }
-}
+//package kitchenpos.eatinorders.domain;
+//
+//import javax.persistence.Column;
+//import javax.persistence.Entity;
+//import javax.persistence.Id;
+//import javax.persistence.Table;
+//import java.util.UUID;
+//
+//@Table(name = "order_table")
+//@Entity
+//public class OrderTable {
+//    @Column(name = "id", columnDefinition = "binary(16)")
+//    @Id
+//    private UUID id;
+//
+//    @Column(name = "name", nullable = false)
+//    private String name;
+//
+//    @Column(name = "number_of_guests", nullable = false)
+//    private int numberOfGuests;
+//
+//    @Column(name = "occupied", nullable = false)
+//    private boolean occupied;
+//
+//    public OrderTable() {
+//    }
+//
+//    public UUID getId() {
+//        return id;
+//    }
+//
+//    public void setId(final UUID id) {
+//        this.id = id;
+//    }
+//
+//    public String getName() {
+//        return name;
+//    }
+//
+//    public void setName(final String name) {
+//        this.name = name;
+//    }
+//
+//    public int getNumberOfGuests() {
+//        return numberOfGuests;
+//    }
+//
+//    public void setNumberOfGuests(final int numberOfGuests) {
+//        this.numberOfGuests = numberOfGuests;
+//    }
+//
+//    public boolean isOccupied() {
+//        return occupied;
+//    }
+//
+//    public void setOccupied(final boolean occupied) {
+//        this.occupied = occupied;
+//    }
+//}
