+++ src/main/java/kitchenpos/eatinorders/tobe/domain/OrderLineItem.java
@@ -2,16 +2,48 @@ package kitchenpos.eatinorders.tobe.domain;
 
 import kitchenpos.menus.tobe.domain.Menu;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import java.math.BigDecimal;
+import java.util.UUID;
+
+@Table(name = "order_line_item")
+@Entity
 public class OrderLineItem {
-    private final Menu menu;
-    private final long quantity;
+    @Column(name = "seq")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    private Long seq;
+    private Menu menu;
+    private long quantity;
+    private UUID menuId;
+        private BigDecimal price;
 
     public OrderLineItem(final Menu menu, final long quantity) {
         this.menu = menu;
         this.quantity = quantity;
     }
 
+    public OrderLineItem() {
+    }
+
+    public Menu getMenu() {
+        return menu;
+    }
+
     public long getQuantity() {
         return quantity;
+    }
+
+    public UUID getMenuId() {
+        return menuId;
+    }
+
+    public BigDecimal getPrice() {
+        return price;
     }
 }
