+++ src/test/java/kitchenpos/eatinorders/application/InMemoryOrderTableRepository.java
@@ -2,6 +2,7 @@ package kitchenpos.eatinorders.application;
 
 import kitchenpos.eatinorders.domain.OrderTable;
 import kitchenpos.eatinorders.domain.OrderTableRepository;
+import kitchenpos.eatinorders.tobe.domain.OrderTable;
 
 import java.util.*;
